+++ include/managers/ResourceFactory.hpp
