+++ tests/resources/ResourceTestAccess.hpp
@@ -1,0 +1,10 @@
+// Test helper for resource system reset. Not part of public API.
+#ifndef TESTS_RESOURCES_RESOURCE_TEST_ACCESS_HPP
+#define TESTS_RESOURCES_RESOURCE_TEST_ACCESS_HPP
+
+struct ResourceTestAccess {
+    static void resetFactory();
+};
+
+#endif // TESTS_RESOURCES_RESOURCE_TEST_ACCESS_HPP
+
